Fix format string bug in mxmotor 'show ver'.
git-svn-id: http://svn.csrri.iit.edu/mx/trunk@4422 42547f41-83fc-0310-9aef-9668e019ab9e

+++ motor/mshow.c
@@ -1233,7 +1233,7 @@ motor_show_version( void )
 /*-------------------------------------------------------------------------*/
 
 	fprintf( output, "Process ID: %lu, Wordsize: %lu\n",
-		mx_process_id(), MX_WORDSIZE );
+		mx_process_id(), (unsigned long) MX_WORDSIZE );
 
 /*-------------------------------------------------------------------------*/
